Remove branching logic from DuelPlayGameRiverpod
